automation/requirements.txt
openai==1.40.2
anthropic==0.34.2
httpx==0.27.0
requests==2.32.3
pydantic==2.8.2
pytest==8.2.2
python-dotenv==1.0.0